requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==2.0.12
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-resources==5.4.0
openai==1.35.14
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
ruff==0.5.2
sniffio==1.3.1
tqdm==4.64.1
typing_extensions==4.12.2
urllib3==1.26.19
zipp==3.6.0
